# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: WhatsApp Store Builder API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all products for the authenticated user's store
      security:
        - bearerAuth: []
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: inStock
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a new product
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "403":
          description: Plan limit reached
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LimitReachedError"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a product by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    put:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductBody"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /products/categories:
    get:
      operationId: listCategories
      tags: [products]
      summary: List all product categories for the store
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string

  # Order endpoints
...
  /referral/me:
    get:
      operationId: getMyReferral
      tags: [referral]
      summary: Get the current user's referral info
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Referral info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReferralInfo"
  /referral/apply:
    post:
      operationId: applyReferralCode
      tags: [referral]
      summary: Apply a referral code
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplyReferralBody"
      responses:
        "200":
          description: Referral applied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReferralInfo"

  # Admin endpoints
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    LimitReachedError:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
          enum: [PRODUCT_LIMIT_REACHED, ORDER_LIMIT_REACHED]
        limit:
          type: integer
        current:
          type: integer
        planName:
          type: string
      required: [error, code, limit, current, planName]
...
    Product:
      type: object
      properties:
        id:
          type: integer
        storeId:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        price:
          type: number
        imageUrl:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
        stock:
          type: integer
          nullable: true
        variants:
          type: array
          items:
            $ref: "#/components/schemas/ProductVariant"
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - storeId
        - name
        - price
        - isActive
        - variants
        - createdAt
    ProductVariant:
      type: object
      properties:
        name:
          type: string
        options:
          type: array
          items:
            type: string
      required: [name, options]
    CreateProductBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        price:
          type: number
        imageUrl:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
        stock:
          type: integer
          nullable: true
        variants:
          type: array
          items:
            $ref: "#/components/schemas/ProductVariant"
        isActive:
          type: boolean
          default: true
      required: [name, price]
    UpdateProductBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        price:
          type: number
        imageUrl:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
        stock:
          type: integer
          nullable: true
        variants:
          type: array
          items:
            $ref: "#/components/schemas/ProductVariant"
        isActive:
          type: boolean
...
    ReferralInfo:
      type: object
      properties:
        referralCode:
          type: string
        referralLink:
          type: string
        referredCount:
          type: integer
        bonusOrdersEarned:
          type: integer
      required: [referralCode, referralLink, referredCount, bonusOrdersEarned]
    ApplyReferralBody:
      type: object
      properties:
        code:
          type: string
      required: [code]
...