# lib/api-zod/src/generated/types/limitReachedError.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * WhatsApp Store Builder API
 * OpenAPI spec version: 0.1.0
 */
import type { LimitReachedErrorCode } from "./limitReachedErrorCode";

export interface LimitReachedError {
  error: string;
  code: LimitReachedErrorCode;
  limit: number;
  current: number;
  planName: string;
}